Create 1.1 checkerboardKarel
My first attempt for the first assignment.  

Karel starts at the corner of 1st avenue and 1st street facing East.  This program runs a loop of two conditions for odd number row and even number row.  The odd number row always start by putting a beeper.  Then, it continues creating a checkered board pattern until reaching the wall.  It then goes back to the beginning and move up to the next row. While the even number row, Karel starts by moving to the next corner and drop a beeper.  Then, it continues creating a checkered board pattern and goes back to the beginning and move up to the next row just like in the odd number row.

The program was tested on all the world given for the assignment.
	 */